requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Cython==0.26
Django==1.11.3
docutils==0.13.1
gunicorn==19.7.1
idna==2.5
Kivy==1.10.0
Kivy-Garden==0.1.4
pyephem==3.7.6.0
pygame==1.9.3
Pygments==2.2.0
pyserial==3.4
pytz==2017.2
requests==2.18.2
serialenum==0.3
tabulate==0.8.1
urllib3==1.22
